refactor: App.tsxからカスタムフックを抽出
- useGameClock: 時計状態・250ms更新・表示用投影
- useGameWebSocket: WebSocket接続・再接続ロジック
- useBotTurn: Bot着手のsetTimeout処理
